REAL AUDIO FOR EXPLOSIONS
this was a huge deal.

0. finding the sound on opengameart. thank ZEUS for that website

1. i had to figure out how to make ffmpeg convert a 24KB file into
something much smaller to fit into 13KB. even 10KB was probably too big
because i want to add extra code. it was always too big using Ogg
Vorbis, so i tried Ogg Opus, but that doesn't even play, so then i fell
back to mp3.

2. converting mp3s to data-uris is actually not fun. i might write a bin
to do it if i have to do this a lot.

3. if you have hecka Audio streams, you're screwed. if the game plays a
total of 1000 explosion sounds, you can't make 1000 audio streams. so
you have to reuse them. i tried to max it out (in a reasonable way,
making a sound a few times a second) and it made 9 streams max. i was
happy with that.

4. because it's mp3, this doesn't work on Firefox (unless you're on
Windows, according to Wikipedia) but it should work elsewhere.

phew. i hate sound